refactor(core): member-service ResponseDto 기반, core로 분리
