+++ core/src/main/java/org/hexagon/core/dto/Empty.java
@@ -1,0 +1,14 @@
+package org.hexagon.core.dto;
+
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+
+@JsonSerialize
+public class Empty {
+    private static final Empty INSTANCE = new Empty();
+
+    private Empty() {}
+
+    public static Empty getInstance() {
+        return INSTANCE;
+    }
+}
